+++ lumen/ai/__init__.py
@@ -2,8 +2,8 @@ import panel as pn
 
 from . import agents, embeddings, llm  # noqa
 from .agents import Analysis  # noqa
+from .coordinator import Coordinator, DependencyResolver, Planner  # noqa
 from .memory import memory  # noqa
-from .orchestrator import Orchestrator, Planner  # noqa
 from .ui import ChatUI, ExplorerUI  # noqa
 
 pn.chat.message.DEFAULT_AVATARS.update({
